Configuration::set, resolve ambiguity HACKish
Added a type coercion for the .count() method that returns QSizeType
There might be a better / safer way but this will do for now.

+++ gui/customactions.cpp
@@ -107,7 +107,8 @@ void CustomActions::set(QList<Command> cmds)
         cfg.removeGroup(metaObject()->className());
         if (!commands.isEmpty()) {
             cfg.beginGroup(metaObject()->className());
-            cfg.set("count", commands.count());
+            // HACK This call is overloaded and ambiguous, QSizeType instead?
+            cfg.set("count", (int) commands.count());
             for (int i=0; i<commands.count(); ++i) {
                 cfg.set(QString::number(i)+QLatin1String("_name"), commands[i].name);
                 cfg.set(QString::number(i)+QLatin1String("_cmd"), commands[i].cmd);
